Add an example
